Add files via upload
The code for counting all directed graphlets of size 2 and 3. The command line Java program takes a file that lists input network file names and a threshold t for input. All the graphlets with counts greater than or equal to t are reported in separate output files for each input network.